requirements.txt
Django==2.2
psycopg2
gunicorn
django-boogie~=0.11.1
requests
